+++ CHANGELOG.md
@@ -6,6 +6,29 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.2.0] - 2024-11-27
+
+### Added
+
+- Information for Get-Help.
+- Examples.
+- Validate parameters for all functions.
+- Parameters for `Add-AdaptiveCardRichTextBlock`:
+  - `SelectAction`
+  - `Url`
+  - `FontType`
+  - `Highlight`
+  - `IsSubtle`
+  - `Italic`
+  - `Strikethrough`
+  - `Underline`
+- Parameters for `Add-AdaptiveCardTextBlock`:
+  - `FontType`
+  - `HAlign`
+  - `IsSubtle`
+  - `MaxLines`
+  - `Wrap`
+
 ## [1.1.0] - 2024-11-24
 
 ### Added